LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/policy-inherited-correctly-by-plznavigate-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/support/fail.html because it does not appear in the frame-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

NOTRUN iframe still inherits correct CSP
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
